[aml] move AMLUtils.cpp to utils dir and conditional build on amcodec or amlplayer set in configure
