+++ configure.in
@@ -2419,7 +2419,8 @@ OUTPUT_FILES="Makefile \
     xbmc/peripherals/devices/Makefile \
     xbmc/android/activity/Makefile \
     xbmc/android/loader/Makefile \
-    xbmc/main/Makefile"
+    xbmc/main/Makefile \
+    xbmc/utils/Makefile"
 
 if test "$use_skin_touched" = "yes"; then
 OUTPUT_FILES="$OUTPUT_FILES addons/skin.touched/media/Makefile"
