+++ tools/amlogic/depends/xbmc/xbmc.mk
@@ -9,7 +9,7 @@ $(SOURCE)/xbmc.bin: $(SOURCE)/Makefile
 $(SOURCE)/Makefile: $(SOURCE)/configure
 	cd $(SOURCE); $(TARGET_CONFIGURE_OPTS) $(TARGET_CONFIGURE_ARGS) $(XBMC_CONF_ENV) \
         ./configure --target=$(GNU_TARGET_NAME) --host=$(GNU_TARGET_NAME) --build=$(GNU_HOST_NAME) \
-        --prefix=/usr --exec-prefix=/usr --sysconfdir=/etc $(XBMC_CONF_OPT) --disable-external-ffmpeg --enable-codec=amcodec
+        --prefix=/usr --exec-prefix=/usr --sysconfdir=/etc $(XBMC_CONF_OPT) --disable-external-ffmpeg
 
 $(SOURCE)/configure: $(BUILDROOT)/package/thirdparty/xbmc/xbmc.mk  $(GCC)
 	cd $(SOURCE); ./bootstrap
