+++ xbmc/utils/AMLUtils.h
