--- xbmc/utils/Makefile
@@ -1,79 +1,0 @@
-SRCS=AlarmClock.cpp \
-     AliasShortcutUtils.cpp \
-     Archive.cpp \
-     AsyncFileCopy.cpp \
-     AutoPtrHandle.cpp \
-     Base64.cpp \
-     BitstreamConverter.cpp \
-     BitstreamStats.cpp \
-     CharsetConverter.cpp \
-     CPUInfo.cpp \
-     Crc32.cpp \
-     CryptThreading.cpp \
-     DatabaseUtils.cpp \
-     DownloadQueue.cpp \
-     DownloadQueueManager.cpp \
-     EndianSwap.cpp \
-     EdenVideoArtUpdater.cpp \
-     Fanart.cpp \
-     fastmemcpy.c \
-     fastmemcpy-arm.S \
-     FileOperationJob.cpp \
-     FileUtils.cpp \
-     fstrcmp.c \
-     fft.cpp \
-     GLUtils.cpp \
-     GroupUtils.cpp \
-     HTMLTable.cpp \
-     HTMLUtil.cpp \
-     HttpHeader.cpp \
-     HttpParser.cpp \
-     HttpResponse.cpp \
-     InfoLoader.cpp \
-     JobManager.cpp \
-     JSONVariantParser.cpp \
-     JSONVariantWriter.cpp \
-     LabelFormatter.cpp \
-     LangCodeExpander.cpp \
-     LCD.cpp \
-     LCDFactory.cpp \
-     log.cpp \
-     md5.cpp \
-     Observer.cpp \
-     Mime.cpp \
-     PerformanceSample.cpp \
-     PerformanceStats.cpp \
-     POUtils.cpp \
-     RecentlyAddedJob.cpp \
-     RegExp.cpp \
-     RingBuffer.cpp \
-     RssReader.cpp \
-     ScraperParser.cpp \
-     ScraperUrl.cpp \
-     Screenshot.cpp \
-     SeekHandler.cpp \
-     SortUtils.cpp \
-     Splash.cpp \
-     Stopwatch.cpp \
-     StreamDetails.cpp \
-     StreamUtils.cpp \
-     StringUtils.cpp \
-     SystemInfo.cpp \
-     TextSearch.cpp \
-     TimeSmoother.cpp \
-     TimeUtils.cpp \
-     TuxBoxUtil.cpp \
-     URIUtils.cpp \
-     UrlOptions.cpp \
-     Variant.cpp \
-     Vector.cpp \
-     Weather.cpp \
-     XBMCTinyXML.cpp \
-     XMLUtils.cpp \
-     XMLVariantParser.cpp \
-     XMLVariantWriter.cpp \
-
-LIB=utils.a
-
-include ../../Makefile.include
--include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(patsubst %.S,,$(SRCS))))
