+++ xbmc/utils/Makefile.in
@@ -1,0 +1,84 @@
+SRCS += AlarmClock.cpp
+SRCS += AliasShortcutUtils.cpp
+SRCS += Archive.cpp
+SRCS += AsyncFileCopy.cpp
+ifeq (@USE_LIBAMCODEC@,1)
+SRCS += AMLUtils.cpp
+else ifeq (@USE_AMLPLAYER@,1)
+SRCS += AMLUtils.cpp
+endif
+SRCS += AutoPtrHandle.cpp
+SRCS += Base64.cpp
+SRCS += BitstreamConverter.cpp
+SRCS += BitstreamStats.cpp
+SRCS += CharsetConverter.cpp
+SRCS += CPUInfo.cpp
+SRCS += Crc32.cpp
+SRCS += CryptThreading.cpp
+SRCS += DatabaseUtils.cpp
+SRCS += DownloadQueue.cpp
+SRCS += DownloadQueueManager.cpp
+SRCS += EndianSwap.cpp
+SRCS += EdenVideoArtUpdater.cpp
+SRCS += Fanart.cpp
+SRCS += fastmemcpy.c
+SRCS += fastmemcpy-arm.S
+SRCS += FileOperationJob.cpp
+SRCS += FileUtils.cpp
+SRCS += fstrcmp.c
+SRCS += fft.cpp
+SRCS += GLUtils.cpp
+SRCS += GroupUtils.cpp
+SRCS += HTMLTable.cpp
+SRCS += HTMLUtil.cpp
+SRCS += HttpHeader.cpp
+SRCS += HttpParser.cpp
+SRCS += HttpResponse.cpp
+SRCS += InfoLoader.cpp
+SRCS += JobManager.cpp
+SRCS += JSONVariantParser.cpp
+SRCS += JSONVariantWriter.cpp
+SRCS += LabelFormatter.cpp
+SRCS += LangCodeExpander.cpp
+SRCS += LCD.cpp
+SRCS += LCDFactory.cpp
+SRCS += log.cpp
+SRCS += md5.cpp
+SRCS += Observer.cpp
+SRCS += Mime.cpp
+SRCS += PerformanceSample.cpp
+SRCS += PerformanceStats.cpp
+SRCS += POUtils.cpp
+SRCS += RecentlyAddedJob.cpp
+SRCS += RegExp.cpp
+SRCS += RingBuffer.cpp
+SRCS += RssReader.cpp
+SRCS += ScraperParser.cpp
+SRCS += ScraperUrl.cpp
+SRCS += Screenshot.cpp
+SRCS += SeekHandler.cpp
+SRCS += SortUtils.cpp
+SRCS += Splash.cpp
+SRCS += Stopwatch.cpp
+SRCS += StreamDetails.cpp
+SRCS += StreamUtils.cpp
+SRCS += StringUtils.cpp
+SRCS += SystemInfo.cpp
+SRCS += TextSearch.cpp
+SRCS += TimeSmoother.cpp
+SRCS += TimeUtils.cpp
+SRCS += TuxBoxUtil.cpp
+SRCS += URIUtils.cpp
+SRCS += UrlOptions.cpp
+SRCS += Variant.cpp
+SRCS += Vector.cpp
+SRCS += Weather.cpp
+SRCS += XBMCTinyXML.cpp
+SRCS += XMLUtils.cpp
+SRCS += XMLVariantParser.cpp
+SRCS += XMLVariantWriter.cpp
+
+LIB   = utils.a
+
+include @abs_top_srcdir@/Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(patsubst %.S,,$(SRCS))))
